1064 - extracted getFormattedMessages out of compute into message helper. Fixed assocaited calls and tests
